@@ -17,32 +17,32 @@
     "clsx": "^2.1.1",
     "iron-session": "^8.0.4",
     "leaflet": "^1.9.4",
-    "next": "^14.2.18",
-    "open-props": "^1.7.7",
+    "next": "^14.2.22",
+    "open-props": "^1.7.10",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-hook-form": "^7.53.2",
+    "react-hook-form": "^7.54.2",
     "react-leaflet": "^4.2.1",
-    "resolve-accept-language": "^3.1.8",
+    "resolve-accept-language": "^3.1.10",
     "sharp": "^0.33.5",
-    "zod": "^3.23.8",
+    "zod": "^3.24.1",
     "zod-validation-error": "^3.4.0"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.15",
-    "@types/leaflet": "^1.9.14",
-    "@types/node": "^20.17.6",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
+    "@types/leaflet": "^1.9.15",
+    "@types/node": "^20.17.11",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
     "autoprefixer": "^10.4.20",
-    "daisyui": "^4.12.14",
+    "daisyui": "^4.12.23",
     "eslint": "^8.57.1",
     "eslint-config-next": "14.2.3",
     "eslint-config-prettier": "^9.1.0",
     "postcss": "^8.4.49",
-    "prettier": "^3.3.3",
+    "prettier": "^3.4.2",
     "prettier-plugin-tailwindcss": "^0.6.9",
-    "tailwindcss": "^3.4.15",
-    "typescript": "^5.6.3"
+    "tailwindcss": "^3.4.17",
+    "typescript": "^5.7.2"
   }
 }
