@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   printWidth: 160,
   singleQuote: false,
   plugins: ["prettier-plugin-tailwindcss"],
